requirements.txt
tensorflow>=2.16.2
keras>=3.11.3
numpy>=1.26.4
scikit-learn>=1.6.1
Pillow>=11.3.0
PyQt5>=5.15.11
kagglehub>=0.4.1